Flush out .trace docs after rebase
